Move tests for `toggle` into separate file
